+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Yellowbox Snowglobe Changelog
 ## 0.2.2
 ### Fixed
-* SnowType Boolean converter now supports None as well
+* SnowType converter now supports None as well
 ## 0.2.1
 ### Fixed
 * a bug with concurrent sessions